reorganize files a bit
